+++ packages/server/jest.setup.ts
@@ -1,9 +1,10 @@
-import { formData, setLogLevel } from '@appsemble/node-utils';
+import { formData, setFixtureBase, setLogLevel } from '@appsemble/node-utils';
 import { request } from 'axios-test-instance';
 import { toMatchImageSnapshot } from 'jest-image-snapshot';
 
 request.interceptors.request.use(formData);
 
+setFixtureBase(__dirname);
 setLogLevel(0);
 
 expect.extend({ toMatchImageSnapshot });