+++ packages/server/src/controllers/action.ts
@@ -114,7 +114,7 @@ async function handleRequestProxy(
       sub: user.id,
       name: user.name,
       email: user.primaryEmail,
-      email_verified: user.EmailAuthorizations[0].verified,
+      email_verified: Boolean(user.EmailAuthorizations?.[0]?.verified),
     },
   );
   const axiosConfig = formatRequestAction(action, data, (remapper, d) =>